add splitPlacesBlanks
This patch updates the documentation for splitPlaces a bit; it was a bit
ambiguous about what it did with extraneous lengths. Since it turns out to be
useful to resolve the given ambiguity in the way that splitPlaces *doesn't* do,
this patch also adds a new function splitPlacesBlanks that may output lists
with trailing empty lists, but guarantees that the output list is exactly as
long as the input list.