put generic build function back in, with lots of explanatory comments
The 0.1.4.3 release did away with a custom @build@ implementation in
favor of importing one from "GHC.Exts", which was (reportedly) faster
for some applications.  However, in the interest of simplicity and
complete Haskell2010 compliance as @split@ is being included in the
Haskel Platform, version 0.2.1.0 has gone back to defining @build@
manually.  This is in line with @split@'s design philosophy of having
efficiency as a non-goal.